add trainer for classification

+++ tweetnlp/text_classification/readme_template.py
@@ -52,12 +52,17 @@ def get_readme(model_name: str,
     else:
         widgets_str = f'- text: {widget_sample_sentence}\n  example_title: "Example"'
 
-    evaluation_result = None
-    if os.path.exists(metric_file):
-        shutil.copy2(metric_file, os.path.basename(model_name))
-        metric_file = pj(os.path.basename(model_name), os.path.basename(metric_file))
-        with open(metric_file) as f:
+    metric_file_model = pj(os.path.basename(model_name), os.path.basename(metric_file))
+    if not os.path.exists(metric_file_model):
+        if os.path.exists(metric_file):
+            shutil.copy2(metric_file, os.path.basename(model_name))
+        else:
+            metric_file_model = None
+    if metric_file_model is not None:
+        with open(metric_file_model) as f:
             evaluation_result = json.load(f)
+    else:
+        evaluation_result = None
     return f"""---
 datasets:
 - {dataset_type}
